Add PixelOverlay and NexusOverlay to our prebuilts.
The pixel phones now have a way to theme android framework without root.
Lets take advantage of this system by including the PixelOverlay from
the offical pixel images and lets make it easy for user to swwitch back
to the default overlay by adding in our own custom NexusOverlay which is
stock colors.